+++ app/View/Composers/Comments.php
@@ -22,8 +22,8 @@ class Comments extends Composer
     {
         return sprintf(
             /* translators: %1$s is replaced with the number of comments and %2$s with the post title */
-            _nx('%1$s response to &ldquo;%2$s&rdquo;', '%1$s responses to &ldquo;%2$s&rdquo;', get_comments_number(), 'comments title', 'sage'),
-            get_comments_number() === 1 ? _x('One', 'comments title', 'sage') : number_format_i18n(get_comments_number()),
+            _nx('%1$s comment on "%2$s"', '%1$s comments on "%2$s"', get_comments_number(), 'comments title', 'sage'),
+            number_format_i18n((int) get_comments_number()),
             get_the_title()
         );
     }
@@ -40,6 +40,8 @@ class Comments extends Composer
         return wp_list_comments([
             'style' => 'ol',
             'short_ping' => true,
+            'avatar_size' => 48,
+            'callback' => 'App\\mh_comment_list_item',
             'echo' => false,
         ]);
     }
@@ -54,7 +56,7 @@ class Comments extends Composer
         }
 
         return get_previous_comments_link(
-            __('&larr; Older comments', 'sage')
+            __('Older comments', 'sage')
         );
     }
 
@@ -68,7 +70,7 @@ class Comments extends Composer
         }
 
         return get_next_comments_link(
-            __('Newer comments &rarr;', 'sage')
+            __('Newer comments', 'sage')
         );
     }